training_script/ERNIE/requirements.txt
datasets==1.1.3
scikit-learn
scipy
paddle-ernie
